부산대_1조_알빠임_5주차

+++ src/main/java/com/example/team1_be/domain/Group/GroupRepository.java
@@ -2,5 +2,5 @@ package com.example.team1_be.domain.Group;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface GroupRepository extends JpaRepository<Group, Integer> {
+public interface GroupRepository extends JpaRepository<Group, Long> {
 }
