+++ src/main/java/com/example/team1_be/domain/Substitute/SubstituteRepository.java
@@ -2,5 +2,5 @@ package com.example.team1_be.domain.Substitute;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SubstituteRepository extends JpaRepository<Substitute, Integer> {
+public interface SubstituteRepository extends JpaRepository<Substitute, Long> {
 }
